feat: Create dedicated chat page and apply global orange theme
This commit implements two major changes:
1.  **Dedicated Chat Page:**
    - Moved all chat UI and logic from the popup in `layouts/default.vue`
      to a new dedicated page at `/chat`.
    - The old chat popup and its associated button have been removed from
      `layouts/default.vue`.
    - Added a "Chat Support" link to `AppHeader.vue` pointing to the new
      `/chat` page.
    - The chat functionality on the dedicated page (sending/receiving
      messages, suggestions, info modal, scrolling, animations) is
      preserved.

2.  **Global Orange Theme Integration:**
    - Updated the application's global color palette to be orange-based,
      aligning with the chatbot's original prominent colors.
    - Modified `assets/css/main.css` to redefine the HSL values for core
      Shadcn UI CSS variables (`--primary`, `--secondary`, `--accent`,
      `--muted`, `--border`, `--input`, `--ring`, etc.) for both light
      and dark themes.
    - The new theme uses a vibrant orange as the primary color, with
      complementary shades for other theme elements.
    - This change affects the appearance of all Shadcn UI components
      (Buttons, Cards, Inputs, etc.) and themed elements sitewide.

These updates provide a dedicated and improved space for chat interactions
and give the entire application a new, consistent orange-themed visual identity.
