+++ assets/css/main.css
@@ -47,74 +47,95 @@
 }
 
 :root {
-  --background: oklch(1 0 0);
-  --foreground: oklch(0.145 0 0);
-  --card: oklch(1 0 0);
-  --card-foreground: oklch(0.145 0 0);
-  --popover: oklch(1 0 0);
-  --popover-foreground: oklch(0.145 0 0);
-  --primary: oklch(0.205 0 0);
-  --primary-foreground: oklch(0.985 0 0);
-  --secondary: oklch(0.97 0 0);
-  --secondary-foreground: oklch(0.205 0 0);
-  --muted: oklch(0.97 0 0);
-  --muted-foreground: oklch(0.556 0 0);
-  --accent: oklch(0.97 0 0);
-  --accent-foreground: oklch(0.205 0 0);
-  --destructive: oklch(0.577 0.245 27.325);
-  --destructive-foreground: oklch(0.577 0.245 27.325);
-  --border: oklch(0.922 0 0);
-  --input: oklch(0.922 0 0);
-  --ring: oklch(0.708 0 0);
-  --chart-1: oklch(0.646 0.222 41.116);
-  --chart-2: oklch(0.6 0.118 184.704);
-  --chart-3: oklch(0.398 0.07 227.392);
-  --chart-4: oklch(0.828 0.189 84.429);
-  --chart-5: oklch(0.769 0.188 70.08);
-  --radius: 0.625rem;
-  --sidebar: oklch(0.985 0 0);
-  --sidebar-foreground: oklch(0.145 0 0);
-  --sidebar-primary: oklch(0.205 0 0);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.97 0 0);
-  --sidebar-accent-foreground: oklch(0.205 0 0);
-  --sidebar-border: oklch(0.922 0 0);
-  --sidebar-ring: oklch(0.708 0 0);
+  --background: hsl(0 0% 100%); /* White */
+  --foreground: hsl(240 10% 3.9%); /* Neutral 900 */
+  --card: hsl(0 0% 100%); /* White */
+  --card-foreground: hsl(240 10% 3.9%); /* Neutral 900 */
+  --popover: hsl(0 0% 100%); /* White */
+  --popover-foreground: hsl(240 10% 3.9%); /* Neutral 900 */
+
+  --primary: hsl(24.6 95% 53.1%); /* Vibrant Orange (Tailwind Orange 500) */
+  --primary-foreground: hsl(0 0% 100%); /* White */
+
+  --secondary: hsl(30 90% 95%); /* Light Orange (like orange-50) */
+  --secondary-foreground: hsl(24.6 95% 40%); /* Darker Orange (like orange-700) */
+
+  --muted: hsl(240 5% 96.1%); /* Cool Gray (zinc-100) */
+  --muted-foreground: hsl(240 4.8% 43.9%); /* Cool Gray (zinc-600) */
+
+  --accent: hsl(30 90% 95%); /* Light Orange (Same as secondary) */
+  --accent-foreground: hsl(24.6 95% 40%); /* Darker Orange (Same as secondary-foreground) */
+
+  --destructive: hsl(0 84.2% 60.2%); /* Red (rose-500) */
+  --destructive-foreground: hsl(0 0% 100%); /* White */
+
+  --border: hsl(30 80% 85%); /* Light Orange custom for borders */
+  --input: hsl(30 80% 87%); /* Slightly lighter orange custom for input */
+  --ring: hsl(24.6 95% 53.1%); /* Vibrant Orange (Same as primary for focus rings) */
+
+  --radius: 0.5rem; /* Existing radius, keep as is */
+
+  /* Chart colors - can be kept or updated if a new palette is desired */
+  --chart-1: hsl(221 83% 53%); /* Blue */
+  --chart-2: hsl(160 70% 37%); /* Green */
+  --chart-3: hsl(40 89% 50%);  /* Yellow */
+  --chart-4: hsl(300 70% 50%); /* Purple */
+  --chart-5: hsl(0 70% 50%);   /* Red */
+
+  /* Sidebar colors - can be kept or updated if a new palette is desired */
+  --sidebar: hsl(0 0% 100%);
+  --sidebar-foreground: hsl(240 10% 3.9%);
+  --sidebar-primary: hsl(24.6 95% 53.1%); /* Orange for primary elements in sidebar */
+  --sidebar-primary-foreground: hsl(0 0% 100%);
+  --sidebar-accent: hsl(30 90% 95%);
+  --sidebar-accent-foreground: hsl(24.6 95% 40%);
+  --sidebar-border: hsl(30 80% 85%);
+  --sidebar-ring: hsl(24.6 95% 53.1%);
 }
 
 .dark {
-  --background: oklch(0.145 0 0);
-  --foreground: oklch(0.985 0 0);
-  --card: oklch(0.145 0 0);
-  --card-foreground: oklch(0.985 0 0);
-  --popover: oklch(0.145 0 0);
-  --popover-foreground: oklch(0.985 0 0);
-  --primary: oklch(0.985 0 0);
-  --primary-foreground: oklch(0.205 0 0);
-  --secondary: oklch(0.269 0 0);
-  --secondary-foreground: oklch(0.985 0 0);
-  --muted: oklch(0.269 0 0);
-  --muted-foreground: oklch(0.708 0 0);
-  --accent: oklch(0.269 0 0);
-  --accent-foreground: oklch(0.985 0 0);
-  --destructive: oklch(0.396 0.141 25.723);
-  --destructive-foreground: oklch(0.637 0.237 25.331);
-  --border: oklch(0.269 0 0);
-  --input: oklch(0.269 0 0);
-  --ring: oklch(0.439 0 0);
-  --chart-1: oklch(0.488 0.243 264.376);
-  --chart-2: oklch(0.696 0.17 162.48);
-  --chart-3: oklch(0.769 0.188 70.08);
-  --chart-4: oklch(0.627 0.265 303.9);
-  --chart-5: oklch(0.645 0.246 16.439);
-  --sidebar: oklch(0.205 0 0);
-  --sidebar-foreground: oklch(0.985 0 0);
-  --sidebar-primary: oklch(0.488 0.243 264.376);
-  --sidebar-primary-foreground: oklch(0.985 0 0);
-  --sidebar-accent: oklch(0.269 0 0);
-  --sidebar-accent-foreground: oklch(0.985 0 0);
-  --sidebar-border: oklch(0.269 0 0);
-  --sidebar-ring: oklch(0.439 0 0);
+  --background: hsl(20 14.3% 4.1%); /* Very Dark Gray (zinc-950) */
+  --foreground: hsl(0 0% 98%); /* Near White (zinc-50) */
+  --card: hsl(20 14.3% 4.1%); /* Same as background */
+  --card-foreground: hsl(0 0% 98%); /* Same as foreground */
+  --popover: hsl(20 14.3% 4.1%);
+  --popover-foreground: hsl(0 0% 98%);
+
+  --primary: hsl(24.6 85% 58%); /* Brighter/more saturated orange for dark mode */
+  --primary-foreground: hsl(25 95% 15%); /* Very dark orange/brown for text on primary */
+
+  --secondary: hsl(25 80% 15%); /* Dark Muted Orange/Brown */
+  --secondary-foreground: hsl(25 90% 70%); /* Lighter Orange Accent */
+
+  --muted: hsl(240 3.7% 15.9%); /* Dark Gray (zinc-800) */
+  --muted-foreground: hsl(240 5.2% 64.9%); /* Lighter Gray (zinc-400) */
+
+  --accent: hsl(25 80% 20%); /* Slightly more saturated dark muted orange */
+  --accent-foreground: hsl(25 90% 75%); /* Brighter Orange Accent */
+
+  --destructive: hsl(0 72.2% 50.6%); /* Slightly less bright red */
+  --destructive-foreground: hsl(0 0% 98%); /* Near white */
+
+  --border: hsl(240 3.7% 20%); /* Dark Gray (slightly lighter than zinc-800) */
+  --input: hsl(240 3.7% 20%); /* Same as border */
+  --ring: hsl(24.6 85% 58%); /* Primary orange for dark mode focus rings */
+
+  /* Chart colors - can be adjusted for dark mode if needed */
+  --chart-1: hsl(221 73% 63%);
+  --chart-2: hsl(160 60% 47%);
+  --chart-3: hsl(40 79% 60%);
+  --chart-4: hsl(300 60% 60%);
+  --chart-5: hsl(0 60% 60%);
+
+  /* Sidebar colors for dark theme */
+  --sidebar: hsl(20 14.3% 5.1%); /* Slightly off-background */
+  --sidebar-foreground: hsl(0 0% 98%);
+  --sidebar-primary: hsl(24.6 85% 58%);
+  --sidebar-primary-foreground: hsl(25 95% 15%);
+  --sidebar-accent: hsl(25 80% 20%);
+  --sidebar-accent-foreground: hsl(25 90% 75%);
+  --sidebar-border: hsl(240 3.7% 20%);
+  --sidebar-ring: hsl(24.6 85% 58%);
 }
 
 @layer base {
